+++ tests/compile-fail/ClrNameConstructorReserved.kt
@@ -1,0 +1,6 @@
+class ClrNameConstructorReserved {
+    @kotlin.clr.ClrName(".ctor")
+    fun ordinaryMethod(): Unit = Unit
+}
+
+fun main() = ClrNameConstructorReserved().ordinaryMethod()