+++ tests/compile-fail/ClrNameTypeInitializerReserved.kt
@@ -1,0 +1,4 @@
+@kotlin.clr.ClrName(".cctor")
+private fun ordinaryStaticMethod(): Unit = Unit
+
+fun main() = ordinaryStaticMethod()